requirements.txt
requests>=2.28.0
selenium>=4.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=0.20.0
tqdm>=4.64.0
urllib3>=1.26.0
webdriver-manager>=3.8.0
fake-useragent>=1.1.0
retry>=0.9.2
logging>=0.4.9
pandas>=1.3.0
cryptography>=38.0.0 